Moved Sparrow header files from public to project.
When archiving a project, all the header files, including those of
linked projects, seem need to be in the "Project" section of "Copy
Headers" section in "Build Phases".

If they aren't, XCode will complain that the app is not
"a single–bundle application or contains multiple products" when you
archive it and Share/Validate it.