SPTextFieldAutoSize
This makes it possible for SPTextFields to automatically resize themselves,
rather than relying on the user to specify a fixed width and height.

There are two auto-size modes: SPAutoSizeSingleLine (which lays out
text on a single line regardless of linebreaks); and SPAutoSizeMultiline.

SPTextField.autoSizeMaxWidth and autoSizeMaxHeight allow control over the bounds
of auto-sized text fields.